backend/pyproject.toml
[project]
name = "stock-analyzer-backend"
version = "0.1.0"
description = "Stock Analyzer — FastAPI Backend"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "python-jose[cryptography]>=3.3",
    "passlib[bcrypt]>=1.7",
    "python-multipart>=0.0.9",
    "sqlalchemy>=2.0",
    "aiosqlite>=0.20",
    "yfinance>=0.2",
    "pandas>=2.0",
    "numpy>=1.24",
    "ta>=0.11",
    "httpx>=0.27",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "httpx>=0.27",
]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.backends._legacy:_Backend"
